Remove ToolSettingsOptionsViewModel in place of ToolSettingsViewModel
