+++ godtools/App/Features/ToolSettings/Presentation/ToolSettings/ToolSettingsViewModel.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import SwiftUI
 import GodToolsToolParser
 import Combine
 
@@ -17,12 +18,18 @@ class ToolSettingsViewModel: ObservableObject {
     private let parallelLanguageSubject: CurrentValueSubject<LanguageModel?, Never>
     private let trainingTipsEnabled: Bool
     private let shareables: [Shareable]
+    private let trainingTipsEnabledSubject: CurrentValueSubject<Bool, Never> = CurrentValueSubject(false)
     
+    private var trainingTipsCancellable: AnyCancellable?
     private var primaryLanguageCancellable: AnyCancellable?
     private var parallelLanguageCancellable: AnyCancellable?
     
     private weak var flowDelegate: FlowDelegate?
     
+    @Published var shareLinkTitle: String = ""
+    @Published var screenShareTitle: String = ""
+    @Published var trainingTipsIcon: SwiftUI.Image = Image("")
+    @Published var trainingTipsTitle: String = ""
     @Published var chooseLanguageTitle: String = ""
     @Published var primaryLanguageTitle: String = ""
     @Published var parallelLanguageTitle: String = ""
@@ -37,6 +44,16 @@ class ToolSettingsViewModel: ObservableObject {
         self.parallelLanguageSubject = parallelLanguageSubject
         self.trainingTipsEnabled = trainingTipsEnabled
         self.shareables = shareables
+        
+        shareLinkTitle = localizationServices.stringForMainBundle(key: "toolSettings.option.shareLink.title")
+        screenShareTitle = localizationServices.stringForMainBundle(key: "toolSettings.option.screenShare.title")
+        
+        trainingTipsEnabledSubject.send(trainingTipsEnabled)
+        
+        trainingTipsCancellable = trainingTipsEnabledSubject.sink { [weak self] (trainingTipsEnabled: Bool) in
+            self?.trainingTipsTitle = trainingTipsEnabled ? localizationServices.stringForMainBundle(key: "toolSettings.option.trainingTips.hide.title") : localizationServices.stringForMainBundle(key: "toolSettings.option.trainingTips.show.title")
+            self?.trainingTipsIcon = trainingTipsEnabled ? Image(ImageCatalog.toolSettingsOptionHideTips.name) : Image(ImageCatalog.toolSettingsOptionTrainingTips.name)
+        }
         chooseLanguageTitle = localizationServices.stringForMainBundle(key: "toolSettings.chooseLanguage.title")
         hidesShareables = shareables.isEmpty
         
@@ -77,20 +94,6 @@ class ToolSettingsViewModel: ObservableObject {
         )
     }
     
-    func getOptionsViewModel() -> BaseToolSettingsOptionsViewModel {
-        
-        guard let flowDelegate = flowDelegate else {
-            assertionFailure("Failed to instantiate viewModel, flowDelegate should not be nil.")
-            return BaseToolSettingsOptionsViewModel()
-        }
-        
-        return ToolSettingsOptionsViewModel(
-            flowDelegate: flowDelegate,
-            localizationServices: localizationServices,
-            trainingTipsEnabled: trainingTipsEnabled
-        )
-    }
-    
     func getShareablesViewModel() -> BaseToolSettingsShareablesViewModel {
         
         guard let flowDelegate = flowDelegate else {
@@ -108,6 +111,25 @@ class ToolSettingsViewModel: ObservableObject {
     
     private func getTranslatedLanguageName(language: LanguageModel) -> String {
         return LanguageViewModel(language: language, localizationServices: localizationServices).translatedLanguageName
+    }
+    
+    func shareLinkTapped() {
+        
+        flowDelegate?.navigate(step: .shareLinkTappedFromToolSettings)
+    }
+    
+    func screenShareTapped() {
+        
+        flowDelegate?.navigate(step: .screenShareTappedFromToolSettings)
+    }
+    
+    func trainingTipsTapped() {
+
+        trainingTipsEnabledSubject.send(!trainingTipsEnabledSubject.value)
+                
+        let step: FlowStep = trainingTipsEnabledSubject.value ? .enableTrainingTipsTappedFromToolSettings : .disableTrainingTipsTappedFromToolSettings
+        
+        flowDelegate?.navigate(step: step)
     }
     
     func primaryLanguageTapped() {
