--- godtools/App/Features/ToolSettings/Presentation/ToolSettings/Views/ToolSettingsOptions/BaseToolSettingsOptionsViewModel.swift
@@ -1,23 +1,0 @@
-//
-//  BaseToolSettingsOptionsViewModel.swift
-//  godtools
-//
-//  Created by Levi Eggert on 5/13/22.
-//  Copyright © 2022 Cru. All rights reserved.
-//
-
-import SwiftUI
-
-class BaseToolSettingsOptionsViewModel: ObservableObject {
-    
-    @Published var shareLinkTitle: String = ""
-    @Published var screenShareTitle: String = ""
-    @Published var trainingTipsIcon: Image = Image("")
-    @Published var trainingTipsTitle: String = ""
-    
-    init() {}
-    
-    func shareLinkTapped() {}
-    func screenShareTapped() {}
-    func trainingTipsTapped() {}
-}
