--- godtools/App/Features/ToolSettings/Presentation/ToolSettings/Views/ToolSettingsOptions/Views/ToolSettingsOptionsItemViewModel.swift
@@ -1,12 +1,0 @@
-//
-//  ToolSettingsOptionsItemViewModel.swift
-//  ToolSettings
-//
-//  Created by Levi Eggert on 5/11/22.
-//
-
-import Foundation
-
-class ToolSettingsOptionsItemViewModel {
-    
-}
